pyproject.toml
[project]
name = "botcalendar2"
version = "0.1.0"
description = "Telegram bot with calendar for notes"
authors = [
    {name = "Clavker",email = "Clavker.Fust@gmail.com"}
]
requires-python = ">=3.12,<3.15"
dependencies = [
    "aiogram (>=3.25.0,<4.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "django (>=6.0.3,<7.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-django (>=4.12.0,<5.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)"
]
